+++ news-bundle/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Contao news bundle change log
 
+### DEV
+
+ * Trigger all the callbacks in the toggleVisibility() methods (see contao/core-bundle#756).
+
 ### 4.4.0-beta1 (2017-05-05)
 
  * Use the file meta data by default when adding an image (see contao/core-bundle#807).